Fix build with geos>=3.3.0
Fix "call of overloaded 'create(int, int)' is ambiguous" build errors
with new geos-3.3.0 by explicitely specifying argument types

See http://trac.openstreetmap.org/ticket/3299

Patch by AMDmi3 ( https://github.com/AMDmi3/osm2pgsql/commit/45f427132e0d9669f9ad4e4ff1dc418a2603f6b7#diff-0 )

+++ build_geometry.cpp
@@ -71,7 +71,7 @@ static std::vector<double> areas;
 
 char *get_wkt_simple(osmNode *nodes, int count, int polygon) {
     GeometryFactory gf;
-    std::auto_ptr<CoordinateSequence> coords(gf.getCoordinateSequenceFactory()->create(0, 2));
+    std::auto_ptr<CoordinateSequence> coords(gf.getCoordinateSequenceFactory()->create((size_t)0, (size_t)2));
 
     try
     {
@@ -113,7 +113,7 @@ char *get_wkt_simple(osmNode *nodes, int count, int polygon) {
 
 size_t get_wkt_split(osmNode *nodes, int count, int polygon, double split_at) {
     GeometryFactory gf;
-    std::auto_ptr<CoordinateSequence> coords(gf.getCoordinateSequenceFactory()->create(0, 2));
+    std::auto_ptr<CoordinateSequence> coords(gf.getCoordinateSequenceFactory()->create((size_t)0, (size_t)2));
     double area;
     WKTWriter wktw;
     size_t wkt_size = 0;
@@ -143,7 +143,7 @@ size_t get_wkt_split(osmNode *nodes, int count, int polygon, double split_at) {
 
             double distance = 0;
             std::auto_ptr<CoordinateSequence> segment;
-            segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create(0, 2));
+            segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create((size_t)0, (size_t)2));
             segment->add(coords->getAt(0));
             for(unsigned i=1; i<coords->getSize(); i++) {
                 segment->add(coords->getAt(i));
@@ -155,7 +155,7 @@ size_t get_wkt_split(osmNode *nodes, int count, int polygon, double split_at) {
                     areas.push_back(0);
                     wkt_size++;
                     distance=0;
-                    segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create(0, 2));
+                    segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create((size_t)0, (size_t)2));
                     segment->add(coords->getAt(i));
                 }
             }
@@ -309,7 +309,7 @@ size_t build_geometry(osmid_t osm_id, struct osmNode **xnodes, int *xcount, int 
     try
     {
         for (int c=0; xnodes[c]; c++) {
-            std::auto_ptr<CoordinateSequence> coords(gf.getCoordinateSequenceFactory()->create(0, 2));
+            std::auto_ptr<CoordinateSequence> coords(gf.getCoordinateSequenceFactory()->create((size_t)0, (size_t)2));
             for (int i = 0; i < xcount[c]; i++) {
                 struct osmNode *nodes = xnodes[c];
                 Coordinate c;
@@ -358,7 +358,7 @@ size_t build_geometry(osmid_t osm_id, struct osmNode **xnodes, int *xcount, int 
                         //std::cerr << "polygon(" << osm_id << ") is no good: points(" << pline->getNumPoints() << "), closed(" << pline->isClosed() << "). " << writer.write(pline.get()) << std::endl;
                 double distance = 0;
                 std::auto_ptr<CoordinateSequence> segment;
-                segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create(0, 2));
+                segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create((size_t)0, (size_t)2));
                 segment->add(pline->getCoordinateN(0));
                 for(unsigned i=1; i<pline->getNumPoints(); i++) {
                     segment->add(pline->getCoordinateN(i));
@@ -370,7 +370,7 @@ size_t build_geometry(osmid_t osm_id, struct osmNode **xnodes, int *xcount, int 
                         areas.push_back(0);
                         wkt_size++;
                         distance=0;
-                        segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create(0, 2));
+                        segment = std::auto_ptr<CoordinateSequence>(gf.getCoordinateSequenceFactory()->create((size_t)0, (size_t)2));
                         segment->add(pline->getCoordinateN(i));
                     }
                 }
